interspire_analysis/requirements.txt
streamlit
mysql-connector-python
python-dotenv
beautifulsoup4
pandas
crewai
langchain-community
scikit-learn
pydantic
pymysql==1.1.0
psycopg2-binary
pymysql
tabulate==0.9.0